tests(*) split e2e tests to folders
