+++ test/e2e/deploy/foo.go
@@ -1,3 +1,3 @@
-package e2e
+package deploy
 
 // Keep this one empty here to make go test happy that there are non-test files in this folder
