+++ test/e2e/deploy/kuma_deploy_universal_transparent_proxy_test.go
@@ -1,4 +1,4 @@
-package e2e_test
+package deploy_test
 
 import (
 	"strings"
