+++ test/e2e/externalservices/testdata/externalservice-https-server.yaml
