+++ test/e2e/healthcheck/foo.go
@@ -1,0 +1,3 @@
+package healthcheck
+
+// Keep this one empty here to make go test happy that there are non-test files in this folder
