+++ test/e2e/healthcheck/healthcheck_universal_test.go
@@ -1,4 +1,4 @@
-package e2e_test
+package healthcheck_test
 
 import (
 	. "github.com/onsi/ginkgo"
