+++ test/e2e/helm/kuma_helm_deploy_multi_apps_test.go
@@ -1,4 +1,4 @@
-package e2e_test
+package helm_test
 
 import (
 	"fmt"
